add basic bitpacking + lut

+++ src/lattice/voxel.rs
@@ -12,23 +12,43 @@
 pub struct Voxel(pub u32);
 
 impl Voxel {
-  pub fn new(rgb: [u8; 3], roughness: u8, emissive: bool, metallic: bool, transparent: bool) -> Self {
-    Voxel(
-        ((rgb[0] as u32) << 24)
-      | ((rgb[1] as u32) << 16)
-      | ((rgb[2] as u32) << 8)
-      | (((roughness & 0xF) as u32) << 4)
-      | ((emissive    as u32) << 3)
-      | ((metallic    as u32) << 2)
-      | ((transparent as u32) << 1)
-    )
-  }
+	pub fn new(
+		rgb: [u8; 3],
+		roughness: u8,
+		emissive: bool,
+		metallic: bool,
+		transparent: bool,
+	) -> Self {
+		Voxel(
+			((rgb[0] as u32) << 24)
+				| ((rgb[1] as u32) << 16)
+				| ((rgb[2] as u32) << 8)
+				| (((roughness & 0xF) as u32) << 4)
+				| ((emissive as u32) << 3)
+				| ((metallic as u32) << 2)
+				| ((transparent as u32) << 1),
+		)
+	}
 
-  pub fn rgb(self)         -> [u8; 3] { [(self.0 >> 24) as u8, (self.0 >> 16) as u8, (self.0 >> 8) as u8] }
-  pub fn roughness(self)   -> u8      { ((self.0 >> 4) & 0xF) as u8 }
-  pub fn emissive(self)    -> bool    { (self.0 >> 3) & 1 != 0 }
-  pub fn metallic(self)    -> bool    { (self.0 >> 2) & 1 != 0 }
-  pub fn transparent(self) -> bool    { (self.0 >> 1) & 1 != 0 }
+	pub fn rgb(self) -> [u8; 3] {
+		[
+			(self.0 >> 24) as u8,
+			(self.0 >> 16) as u8,
+			(self.0 >> 8) as u8,
+		]
+	}
+	pub fn roughness(self) -> u8 {
+		((self.0 >> 4) & 0xF) as u8
+	}
+	pub fn emissive(self) -> bool {
+		(self.0 >> 3) & 1 != 0
+	}
+	pub fn metallic(self) -> bool {
+		(self.0 >> 2) & 1 != 0
+	}
+	pub fn transparent(self) -> bool {
+		(self.0 >> 1) & 1 != 0
+	}
 }
 
 // 256-entry perceptually uniform color palette in OKLab space, precomputed
@@ -36,24 +56,24 @@ impl Voxel {
 // bounding the global voxel LUT to a manageable size.
 #[derive(Clone, Debug, Default)]
 pub struct ColorPalette {
-  pub entries: Vec<[u8; 3]>,
+	pub entries: Vec<[u8; 3]>,
 }
 
 impl ColorPalette {
-  pub fn new() -> Self {
-    Self::default()
-  }
+	pub fn new() -> Self {
+		Self::default()
+	}
 
-  pub fn len(&self) -> usize {
-    self.entries.len()
-  }
+	pub fn len(&self) -> u32 {
+		self.entries.len() as u32
+	}
 
-  pub fn is_empty(&self) -> bool {
-    self.entries.is_empty()
-  }
+	pub fn is_empty(&self) -> bool {
+		self.entries.is_empty()
+	}
 
-  // Returns the index of the nearest palette entry to the given linear RGB.
-  pub fn nearest(&self, rgb: [u8; 3]) -> u8 {
-    todo!()
-  }
+	// Returns the index of the nearest palette entry to the given linear RGB.
+	pub fn nearest(&self, rgb: [u8; 3]) -> u8 {
+		todo!()
+	}
 }
